+++ src/main/java/paletteEditor/PaletteEditor.java
@@ -250,12 +250,18 @@ public class PaletteEditor
         return palette;
     }
 
-    // Returns color scaled to 0-0xf8.
+    // Returns color scaled to 0-0xff.
     private java.awt.Color color(int offset) {
         // gggrrrrr 0bbbbbgg
         int r = (romImage[offset] & 0x1f) << 3;
         int g = ((romImage[offset + 1] & 3) << 6) | ((romImage[offset] & 0xe0) >> 2);
         int b = (romImage[offset + 1] << 1) & 0xf8;
+        r *= 0xff;
+        g *= 0xff;
+        b *= 0xff;
+        r /= 0xf8;
+        g /= 0xf8;
+        b /= 0xf8;
         return new java.awt.Color(r, g, b);
     }